requirements.txt
-i https://pypi.org/simple
asgiref==3.4.1; python_version >= '3.6'
certifi==2021.10.8
charset-normalizer==2.0.7; python_version >= '3'
django==3.2.8
djangorestframework==3.12.4
idna==3.3; python_version >= '3'
pytz==2021.3
requests==2.26.0
sqlparse==0.4.2; python_version >= '3.5'
typing-extensions==3.10.0.2; python_version < '3.8'
urllib3==1.26.7; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4' and python_version < '4'
